// plugins/org.eclipse.osee.web/src/app/ple/messaging/message-tables/lib/types/AddSubMessageDialog.ts
/*********************************************************************
 * Copyright (c) 2021 Boeing
 *
 * This program and the accompanying materials are made
 * available under the terms of the Eclipse Public License 2.0
 * which is available at https://www.eclipse.org/legal/epl-2.0/
 *
 * SPDX-License-Identifier: EPL-2.0
 *
 * Contributors:
 *     Boeing - initial API and implementation
 **********************************************************************/

import type { subMessage } from '@osee/messaging/shared/types';

export interface AddSubMessageDialog {
	id: string;
	name: string;
	subMessage: subMessage;
}